+++ CHANGELOG.md
@@ -7,6 +7,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Fixed
+* csharp: System must be imported ([#13]).
+
 ### Changed
 * Parse match blocks better by ignoring end condition for nested groups ([#13]).